+++ Automate/Framework/Machines/Objects/SolarPanelMachine.cs
@@ -4,9 +4,9 @@ using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
 {
-    /// <summary>A tapper that accepts input and provides output.</summary>
-    /// <remarks>Derived from <see cref="SObject.performDropDownAction"/> and <see cref="SObject.checkForAction"/> (search for 'Worm Bin').</remarks>
-    internal class WormBinMachine : GenericObjectMachine<SObject>
+    /// <summary>A solar panel that provides output.</summary>
+    /// <remarks>Derived from <see cref="SObject.performDropDownAction"/> (search for 'Solar Panel') and <see cref="SObject.checkForAction"/> (search for 'case 231').</remarks>
+    internal class SolarPanelMachine : GenericObjectMachine<SObject>
     {
         /*********
         ** Public methods
@@ -15,20 +15,13 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <param name="machine">The underlying machine.</param>
         /// <param name="location">The location containing the machine.</param>
         /// <param name="tile">The tile covered by the machine.</param>
-        public WormBinMachine(SObject machine, GameLocation location, Vector2 tile)
+        public SolarPanelMachine(SObject machine, GameLocation location, Vector2 tile)
             : base(machine, location, tile) { }
 
         /// <summary>Get the output item.</summary>
         public override ITrackedStack GetOutput()
         {
-            SObject bin = this.Machine;
-            return new TrackedItem(bin.heldObject.Value, item =>
-            {
-                bin.heldObject.Value = new SObject(685, Game1.random.Next(2, 6));
-                bin.MinutesUntilReady = Utility.CalculateMinutesUntilMorning(Game1.timeOfDay);
-                bin.readyForHarvest.Value = false;
-                bin.showNextIndex.Value = false;
-            });
+            return new TrackedItem(this.Machine.heldObject.Value, onEmpty: this.Reset);
         }
 
         /// <summary>Provide input to the machine.</summary>
@@ -37,6 +30,20 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         public override bool SetInput(IStorage input)
         {
             return false; // no input
+        }
+
+
+        /*********
+        ** Private methods
+        *********/
+        /// <summary>Reset the solar panel.</summary>
+        /// <param name="item">The output item that was taken.</param>
+        private void Reset(Item item)
+        {
+            this.GenericReset(item);
+
+            this.Machine.heldObject.Value = new Object(Vector2.Zero, 787, null, canBeSetDown: false, canBeGrabbed: true, isHoedirt: false, isSpawnedObject: false);
+            this.Machine.MinutesUntilReady = 16800;
         }
     }
 }