+++ Automate/Framework/Machines/Objects/TapperMachine.cs
@@ -1,5 +1,6 @@
 using Microsoft.Xna.Framework;
 using StardewValley;
+using StardewValley.TerrainFeatures;
 using SObject = StardewValley.Object;
 
 namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
@@ -11,8 +12,8 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /*********
         ** Fields
         *********/
-        /// <summary>The tree type.</summary>
-        private readonly int TreeType;
+        /// <summary>The tree being tapped.</summary>
+        private readonly Tree Tree;
 
 
         /*********
@@ -22,11 +23,11 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <param name="machine">The underlying machine.</param>
         /// <param name="location">The location to search.</param>
         /// <param name="tile">The tile covered by the machine.</param>
-        /// <param name="treeType">The tree type being tapped.</param>
-        public TapperMachine(SObject machine, GameLocation location, Vector2 tile, int treeType)
+        /// <param name="tree">The tree being tapped.</param>
+        public TapperMachine(SObject machine, GameLocation location, Vector2 tile, Tree tree)
             : base(machine, location, tile)
         {
-            this.TreeType = treeType;
+            this.Tree = tree;
         }
 
         /// <summary>Get the output item.</summary>
@@ -52,34 +53,9 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         /// <param name="item">The output item that was taken.</param>
         private void Reset(Item item)
         {
-            SObject tapper = this.Machine;
-
-            switch (this.TreeType)
-            {
-                case 1:
-                    tapper.heldObject.Value = new SObject(725, 1);
-                    tapper.MinutesUntilReady = 13000 - Game1.timeOfDay;
-                    break;
-                case 2:
-                    tapper.heldObject.Value = new SObject(724, 1);
-                    tapper.MinutesUntilReady = 16000 - Game1.timeOfDay;
-                    break;
-                case 3:
-                    tapper.heldObject.Value = new SObject(726, 1);
-                    tapper.MinutesUntilReady = 10000 - Game1.timeOfDay;
-                    break;
-                case 7:
-                    tapper.heldObject.Value = new SObject(420, 1);
-                    tapper.MinutesUntilReady = 3000 - Game1.timeOfDay;
-                    if (!Game1.currentSeason.Equals("fall"))
-                    {
-                        tapper.heldObject.Value = new SObject(404, 1);
-                        tapper.MinutesUntilReady = 6000 - Game1.timeOfDay;
-                    }
-                    break;
-            }
-            tapper.heldObject.Value = (SObject)tapper.heldObject.Value.getOne();
-            tapper.readyForHarvest.Value = false;
+            this.Machine.heldObject.Value = null;
+            this.Machine.readyForHarvest.Value = false;
+            this.Tree.UpdateTapperProduct(tapper_instance: this.Machine, previous_object: item as SObject);
         }
     }
 }