+++ LookupAnything/release-notes.md
@@ -9,7 +9,8 @@
   * added island field office lookup;
   * added island Gourmand lookup;
   * added island mermaid puzzle lookup;
-  * added island shrine puzzle lookup.
+  * added island shrine puzzle lookup;
+  * new chest types.
 * Fixed some untranslated text.
 * Fixed NPC lookups not showing social info for some custom mod NPCs. (This affects NPCs with friendship data but `CanSocialize` set to false.)
 * Improved translations. Thanks to spindensity (updated Chinese)!